#    modified:   src/App.tsx #    modified:   src/components/Input/index.tsx #    modified:   src/components/Input/styles.ts #    renamed:    src/pages/Singin.tsx -> src/pages/singIn/Singin.tsx #    renamed:    src/pages/styles.ts -> src/pages/singIn/styles.ts #    new file:   src/pages/transactions/Transactions.tsx
