+++ src/components/Input/index.tsx
@@ -8,7 +8,8 @@ export function Input({icon: Icon, ...rest}: InputProps) {
     return(
         <Container>
            {Icon && <Icon size= {20}/>}
-            <input {...rest} />
+            <input {...rest}
+            />
         </Container>
 
     )
