+++ src/pages/transactions/Transactions.tsx
@@ -1,0 +1,3 @@
+export function Transactions() {
+    return <h1>Transactions</h1>;
+}
